refactor(rpc): move ops handlers to routes/ops.rs (backlog #12 phase 2c)
Next routes slice. Five operator-facing handlers come out:

  - root              (GET /)                  — self-describe
  - health            (GET /health)             — liveness
  - sentrix_status    (GET /sentrix_status)     — structured status
                                                  (still pub, tests import it)
  - metrics           (GET /metrics)            — Prometheus text
  - get_admin_log     (GET /admin/log)          — auth-gated audit log

Shared START_TIME OnceLock lives alongside them — both sentrix_status
and metrics read it for uptime, and create_router still eagerly pins
it at boot so the first handler call doesn't see zero.

Zero route / behaviour change.
